data: move msweb + fix CLI imports
