+++ src/lenskit/cli/data/convert.py
@@ -8,8 +8,6 @@ from pathlib import Path
 
 import click
 
-from lenskit.data import load_amazon_ratings, load_movielens
-from lenskit.data.msweb import load_ms_web
 from lenskit.logging import get_logger
 
 _log = get_logger(__name__)
@@ -41,13 +39,20 @@ def convert(format: str | None, src: list[Path], dst: Path, item_lists: bool = F
             _log.error("no data format specified")
             raise click.UsageError("no data format specified")
         case "movielens":
+            from lenskit.data.sources.movielens import load_movielens
+
             log.info("loading MovieLens data")
             if len(src) != 1:
                 log.error("received %d source paths, MovieLens only takes one", len(src))
+
             data = load_movielens(src[0])
         case "amazon":
+            from lenskit.data.sources.amazon import load_amazon_ratings
+
             data = load_amazon_ratings(*src)
         case "ms-web":
+            from lenskit.data.sources.msweb import load_ms_web
+
             data = load_ms_web(src[0])
         case _:
             raise ValueError(f"unknown data format {format}")
